gif_sim/requirements.txt
pandas
pillow
transformers
numpy>=1.21.0,<2.0.0
sentence-transformers
faiss-gpu
torch
torchvision
torchaudio
fastapi
uvicorn
tqdm
